Safari-611.1.10/LayoutTests/http/tests/media/media-play-stream-chunked-icy-expected.txt


Test ended by:

Play streamed media.
Verify that the 'ended' event fires upon finishing the playback.

EVENT(ended)
END OF TEST
